--- diktat-rules/src/test/resources/test/chapter6/compact_initialization/ApplyInFunctionExtensionsExpected.kt
@@ -1,17 +1,0 @@
-fun String.createPluginConfig() {
-    val pluginConfig = TomlDecoder.decode<T>(
-        serializer(),
-        fakeFileNode,
-        DecoderConf()
-    ).apply {
-    configLocation = this@createPluginConfig.toPath()}
-}
-
-class HttpServer {
-    val defaultPort = "8080"
-    fun foo() {
-        val httpClient = HttpClient("myConnection").apply {
-        url = "http://example.com"
-        port = this@HttpServer.defaultPort}
-    }
-}
