--- diktat-rules/src/test/resources/test/chapter6/compact_initialization/ApplyInFunctionExtensionsTest.kt
@@ -1,17 +1,0 @@
-fun String.createPluginConfig() {
-    val pluginConfig = TomlDecoder.decode<T>(
-        serializer(),
-        fakeFileNode,
-        DecoderConf()
-    )
-    pluginConfig.configLocation = this.toPath()
-}
-
-class HttpServer {
-    val defaultPort = "8080"
-    fun foo() {
-        val httpClient = HttpClient("myConnection")
-        httpClient.url = "http://example.com"
-        httpClient.port = this.defaultPort
-    }
-}
