+++ diktat-rules/src/test/resources/test/chapter6/compact_initialization/ApplyOnStatementsWithThisKeywordExpected.kt
@@ -1,0 +1,14 @@
+fun String.createPluginConfig() {
+    val pluginConfig = TomlDecoder.decode<T>(
+        serializer(),
+        fakeFileNode,
+        DecoderConf()
+    ).apply {
+    prop1 = property1
+    // comment2
+    prop2 = property2}
+    pluginConfig.configLocation = this.toPath()
+    // comment1
+    pluginConfig.configLocation2 = this.toPath()
+}
+
