+++ diktat-rules/src/test/resources/test/chapter6/compact_initialization/ApplyOnStatementsWithThisKeywordTest.kt
@@ -1,0 +1,14 @@
+fun String.createPluginConfig() {
+    val pluginConfig = TomlDecoder.decode<T>(
+        serializer(),
+        fakeFileNode,
+        DecoderConf()
+    )
+    pluginConfig.configLocation = this.toPath()
+    pluginConfig.prop1 = property1
+    // comment1
+    pluginConfig.configLocation2 = this.toPath()
+    // comment2
+    pluginConfig.prop2 = property2
+}
+
